+++ CHANGELOG.rst
@@ -6,6 +6,14 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.1.0/>`__,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`__.
 
+`Unreleased`_
+-------------
+
+Added
+~~~~~
+- Add ``RestClient.search`` method for performing package searches using the V3 API
+
+
 `1.1.0`_ - 2026-01-28
 ---------------------